requirements.txt
aioredis==1.3.1
amqp==5.0.7
-e git+https://github.com/Zakhrov/PyAPNs2.git@5e4a938cd2e24249f6345aac27ba783defa5a63f#egg=apns2
arrow==1.2.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
billiard==3.6.4.0
boto3==1.20.24
botocore==1.23.24
CacheControl==0.12.10
cachetools==4.2.4
celery==5.2.1
celery-progress==0.1.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
daphne==3.0.2
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.10
django-appconf==1.0.5
django-cacheops==6.0
django-celery-results==2.2.0
django-cors-headers==3.10.1
django-croppie==1.3
django-db-geventpool==4.0.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-heroku==0.3.1
-e git+https://github.com/matthewwithanm/django-imagekit.git@2106b2e159f0d4a538532caca9654d4b0d8d8700#egg=django_imagekit
django-ipware==4.0.2
django-oauth-toolkit==1.5.0
django-phonenumber-field==6.0.0
django-push-notifications==2.0.0
django-redis==5.2.0
django-s3direct==1.1.8
django-social-share==2.2.1
django-storages==1.12.3
django-structlog==2.2.0
django-timezone-field==4.2.1
djangorestframework==3.13.1
djangorestframework-bulk==0.2.1
djangorestframework-gis==0.17
dnspython==1.16.0
drf-social-oauth2==1.0.9
drf-yasg==1.20.0
et-xmlfile==1.1.0
eventlet==0.30.2
firebase-admin==5.2.0
funcy==1.17
future==0.18.2
geocoder==1.38.1
gevent==21.12.0
google-api-core==2.3.1
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
h11==0.13.0
h2==2.6.2
hiredis==2.0.0
hpack==3.0.0
http-ece==1.1.0
httplib2==0.20.2
httptools==0.3.0
hyper==0.7.0
hyperframe==3.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
jwcrypto==1.0
kombu==5.2.2
MarkupSafe==2.0.1
msgpack==1.0.3
ndg-httpsclient==0.5.1
numpy==1.21.4
oauth2client==4.1.3
oauthlib==3.1.1
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
pep517==0.12.0
phonenumbers==8.12.39
pilkit==2.0
Pillow==8.4.0
pip-tools==6.5.1
prompt-toolkit==3.0.24
proto-plus==1.19.8
protobuf==3.19.1
psycogreen==1.0.2
psycopg2==2.9.2
psycopg2-binary==2.9.2
pusher==3.1.0
pusher-push-notifications==2.0.1
py-firebase-dynamic-links==1.0.1
py-vapid==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-http-client==3.3.4
python3-openid==3.2.0
pytz==2021.3
pywatchman==1.4.1
pywebpush==1.14.0
PyYAML==6.0
ratelim==0.1.6
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
sendgrid==6.9.3
sentry-sdk==1.5.1
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
stripe==2.63.0
structlog==21.5.0
timezonefinder==5.2.0
tomli==2.0.1
twilio==6.50.1
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.17.1
uvloop==0.16.0
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
whitenoise==5.3.0
wrapt==1.13.3
xlrd==2.0.1
XlsxWriter==3.0.3
zope.event==4.5.0
zope.interface==5.4.0
